d20_requirements.txt
prefect>=2.14
prefect-gcp>=0.4
pandas>=1.5
google-cloud-storage>=2.10
google-cloud-bigquery>=3.12
python-dotenv>=1.0